refactor: database.py into models folder
